+++ crates/ruma-common/CHANGELOG.md
@@ -23,7 +23,7 @@ Improvements:
 - Add `InitialStateEvent::{new, to_raw, to_raw_any}`
 - Add a convenience method to construct `RoomEncryptionEventContent` with the recommended defaults.
 - `PushCondition::EventMatch` and `FlattenedJson` now use escaped dotted paths, according to MSC3873
-- Add unstable support for `event_property_is` push condition according to MSC3758
+- Add support for `event_property_is` push condition (MSC3758 / Matrix 1.7)
 - Add unstable support for `event_property_contains` push condition according to MSC3966
 - Add `FullStateEventContent::redact`
 - Add new methods for `RoomPowerLevels`: